Reorganized the company profile components.
