+++ web-app/src/pages/company-profile/articles/salary.tsx
