+++ web-app/src/pages/company-profile/companyRatingsComponent.tsx
