+++ web-app/src/pages/company-profile/sections/overview.tsx
@@ -5,7 +5,7 @@ import {
 	SectionContainer,
 	SectionHeaderContainer,
 	SectionHeaderTitle,
-} from "../../components";
+} from "../components";
 
 const T = translations.legacyTranslationsNeedsRefactor;
