+++ web-app/src/pages/company-profile/tabs/index.ts
@@ -1,7 +1,7 @@
 import OverviewTab from "./overview";
 import ReviewTab from "./reviews";
 import JobTab from "./jobs";
-import { SalaryTab } from "./salaries";
+import SalaryTab from "./salaries";
 import ContactTab from "./contact";
 
 export { OverviewTab, ReviewTab, JobTab, SalaryTab, ContactTab };
