--- web-app/src/pages/company-profile/tabs/jobs/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./jobs";
