--- web-app/src/pages/company-profile/tabs/salaries/index.ts
@@ -1,4 +1,0 @@
-import SalaryTab from "./salaries";
-import SalaryPosting from "./salary-posting";
-
-export { SalaryPosting, SalaryTab };
